+++ argentum-api/src/main/java/net/devium/argentum/rest/model/ProductRangeResponseEager.java
@@ -3,20 +3,20 @@ package net.devium.argentum.rest.model;
 import java.util.List;
 
 public class ProductRangeResponseEager {
-    private int id;
+    private long id;
     private String name;
     private List<ProductResponseMeta> products;
 
     public ProductRangeResponseEager() {
     }
 
-    public ProductRangeResponseEager(int id, String name, List<ProductResponseMeta> products) {
+    public ProductRangeResponseEager(long id, String name, List<ProductResponseMeta> products) {
         this.id = id;
         this.name = name;
         this.products = products;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
